feat: implement permissions
* feat: button experimentation

* fix: technologies

* fix: add looping

* feat: add MCD and MLD

* feat: split api cached

* fix: server fetches

* feat: rename actions services

* feat: clear examples

* feat: permission first attempt

* feat: add safe message

* feat: add has permission for server actions

* feat: complete refactor

* feat: update deps

* fix: tests

+++ services/actions/QuantityAction.ts
@@ -1,79 +1,87 @@
 "use server";
 
+import { requiresSafeMessage } from "@permissions/requiresSafeMessage";
 import QuantityService from "@services/class/QuantityClass";
 import { QuantityCountProps, QuantityCountResponse, QuantityCreateManyProps, QuantityCreateManyResponse, QuantityCreateProps, QuantityCreateResponse, QuantityDeleteManyProps, QuantityDeleteManyResponse, QuantityDeleteProps, QuantityDeleteResponse, QuantityFindFirstProps, QuantityFindFirstResponse, QuantityFindManyProps, QuantityFindManyResponse, QuantityFindUniqueProps, QuantityFindUniqueResponse, QuantityUpdateManyProps, QuantityUpdateManyResponse, QuantityUpdateProps, QuantityUpdateResponse, QuantityUpsertProps, QuantityUpsertResponse } from "@services/types/QuantityType";
 
 // ========== Single mutations ========== //
 
-export const QuantityCreate = async <T extends QuantityCreateProps>(props: T): Promise<QuantityCreateResponse<T>> => {
+export const QuantityCreateAction = async <T extends QuantityCreateProps>(props: T, disableSafeMessage: boolean = false): Promise<QuantityCreateResponse<T>> => {
     try {
+        await requiresSafeMessage(disableSafeMessage, "QuantityCreateAction", "Quantity", "create");
         const { data, error } = await QuantityService.create(props);
         if (!data || error) throw new Error(error);
         return data;
     } catch (error) {
-        throw new Error("QuantityCreate -> " + (error as Error).message);
+        throw new Error("QuantityCreateAction -> " + (error as Error).message);
     }
 };
 
-export const QuantityUpsert = async <T extends QuantityUpsertProps>(props: T): Promise<QuantityUpsertResponse<T>> => {
+export const QuantityUpsertAction = async <T extends QuantityUpsertProps>(props: T, disableSafeMessage: boolean = false): Promise<QuantityUpsertResponse<T>> => {
     try {
+        await requiresSafeMessage(disableSafeMessage, "QuantityUpsertAction", "Quantity", "upsert");
         const { data, error } = await QuantityService.upsert(props);
         if (!data || error) throw new Error(error);
         return data;
     } catch (error) {
-        throw new Error("QuantityUpsert -> " + (error as Error).message);
+        throw new Error("QuantityUpsertAction -> " + (error as Error).message);
     }
 };
 
-export const QuantityUpdate = async <T extends QuantityUpdateProps>(props: T): Promise<QuantityUpdateResponse<T>> => {
+export const QuantityUpdateAction = async <T extends QuantityUpdateProps>(props: T, disableSafeMessage: boolean = false): Promise<QuantityUpdateResponse<T>> => {
     try {
+        await requiresSafeMessage(disableSafeMessage, "QuantityUpdateAction", "Quantity", "update");
         const { data, error } = await QuantityService.update(props);
         if (!data || error) throw new Error(error);
         return data;
     } catch (error) {
-        throw new Error("QuantityUpdate -> " + (error as Error).message);
+        throw new Error("QuantityUpdateAction -> " + (error as Error).message);
     }
 };
 
-export const QuantityDelete = async <T extends QuantityDeleteProps>(props: T): Promise<QuantityDeleteResponse<T>> => {
+export const QuantityDeleteAction = async <T extends QuantityDeleteProps>(props: T, disableSafeMessage: boolean = false): Promise<QuantityDeleteResponse<T>> => {
     try {
+        await requiresSafeMessage(disableSafeMessage, "QuantityDeleteAction", "Quantity", "delete");
         const { data, error } = await QuantityService.delete(props);
         if (!data || error) throw new Error(error);
         return data;
     } catch (error) {
-        throw new Error("QuantityDelete -> " + (error as Error).message);
+        throw new Error("QuantityDeleteAction -> " + (error as Error).message);
     }
 };
 
 // ========== Multiple mutations ========== //
 
-export const QuantityCreateMany = async (props: QuantityCreateManyProps): Promise<QuantityCreateManyResponse> => {
+export const QuantityCreateManyAction = async (props: QuantityCreateManyProps, disableSafeMessage: boolean = false): Promise<QuantityCreateManyResponse> => {
     try {
+        await requiresSafeMessage(disableSafeMessage, "QuantityCreateManyAction", "Quantity", "createMany");
         const { data, error } = await QuantityService.createMany(props);
         if (!data || error) throw new Error(error);
         return data;
     } catch (error) {
-        throw new Error("QuantityCreateMany -> " + (error as Error).message);
+        throw new Error("QuantityCreateManyAction -> " + (error as Error).message);
     }
 };
 
-export const QuantityUpdateMany = async (props: QuantityUpdateManyProps): Promise<QuantityUpdateManyResponse> => {
+export const QuantityUpdateManyAction = async (props: QuantityUpdateManyProps, disableSafeMessage: boolean = false): Promise<QuantityUpdateManyResponse> => {
     try {
+        await requiresSafeMessage(disableSafeMessage, "QuantityUpdateManyAction", "Quantity", "updateMany");
         const { data, error } = await QuantityService.updateMany(props);
         if (!data || error) throw new Error(error);
         return data;
     } catch (error) {
-        throw new Error("QuantityUpdateMany -> " + (error as Error).message);
+        throw new Error("QuantityUpdateManyAction -> " + (error as Error).message);
     }
 };
 
-export const QuantityDeleteMany = async (props: QuantityDeleteManyProps): Promise<QuantityDeleteManyResponse> => {
+export const QuantityDeleteManyAction = async (props: QuantityDeleteManyProps, disableSafeMessage: boolean = false): Promise<QuantityDeleteManyResponse> => {
     try {
+        await requiresSafeMessage(disableSafeMessage, "QuantityDeleteManyAction", "Quantity", "deleteMany");
         const { data, error } = await QuantityService.deleteMany(props);
         if (!data || error) throw new Error(error);
         return data;
     } catch (error) {
-        throw new Error("QuantityDeleteMany -> " + (error as Error).message);
+        throw new Error("QuantityDeleteManyAction -> " + (error as Error).message);
     }
 };
 
@@ -82,45 +90,51 @@ export const QuantityDeleteMany = async (props: QuantityDeleteManyProps): Promis
 /**
  * WARNING: do not use this for fetching data -> use API routes with caching instead
  */
-export const QuantityFindFirst = async <T extends QuantityFindFirstProps>(
-    props: T
+export const QuantityFindFirstAction = async <T extends QuantityFindFirstProps>(
+    props: T,
+    disableSafeMessage: boolean = false
 ): Promise<QuantityFindFirstResponse<T>> => {
     try {
+        await requiresSafeMessage(disableSafeMessage, "QuantityFindFirstAction", "Quantity", "findFirst");
         const { data, error } = await QuantityService.findFirst(props);
         if (error) throw new Error(error);
         return data ?? null;
     } catch (error) {
-        throw new Error("QuantityFindFirst -> " + (error as Error).message);
+        throw new Error("QuantityFindFirstAction -> " + (error as Error).message);
     }
 };
 
 /**
  * WARNING: do not use this for fetching data -> use API routes with caching instead
  */
-export const QuantityFindUnique = async <T extends QuantityFindUniqueProps>(
-    props: T
+export const QuantityFindUniqueAction = async <T extends QuantityFindUniqueProps>(
+    props: T,
+    disableSafeMessage: boolean = false
 ): Promise<QuantityFindUniqueResponse<T>> => {
     try {
+        await requiresSafeMessage(disableSafeMessage, "QuantityFindUniqueAction", "Quantity", "findUnique");
         const { data, error } = await QuantityService.findUnique(props);
         if (error) throw new Error(error);
         return data ?? null;
     } catch (error) {
-        throw new Error("QuantityFindUnique -> " + (error as Error).message);
+        throw new Error("QuantityFindUniqueAction -> " + (error as Error).message);
     }
 };
 
 /**
  * WARNING: do not use this for fetching data -> use API routes with caching instead
  */
-export const QuantityFindMany = async <T extends QuantityFindManyProps>(
-    props: T
+export const QuantityFindManyAction = async <T extends QuantityFindManyProps>(
+    props: T,
+    disableSafeMessage: boolean = false
 ): Promise<QuantityFindManyResponse<T>> => {
     try {
+        await requiresSafeMessage(disableSafeMessage, "QuantityFindManyAction", "Quantity", "findMany");
         const { data, error } = await QuantityService.findMany(props);
         if (!data || error) throw new Error(error);
         return data;
     } catch (error) {
-        throw new Error("QuantityFindMany -> " + (error as Error).message);
+        throw new Error("QuantityFindManyAction -> " + (error as Error).message);
     }
 };
 
@@ -129,12 +143,13 @@ export const QuantityFindMany = async <T extends QuantityFindManyProps>(
 /**
  * WARNING: do not use this for fetching data -> use API routes with caching instead
  */
-export const QuantityCount = async (props: QuantityCountProps): Promise<QuantityCountResponse> => {
+export const QuantityCountAction = async (props: QuantityCountProps, disableSafeMessage: boolean = false): Promise<QuantityCountResponse> => {
     try {
+        await requiresSafeMessage(disableSafeMessage, "QuantityCountAction", "Quantity", "count");
         const { data, error } = await QuantityService.count(props);
         if (!data || error) throw new Error(error);
         return data;
     } catch (error) {
-        throw new Error("QuantityCount -> " + (error as Error).message);
+        throw new Error("QuantityCountAction -> " + (error as Error).message);
     }
 };
